Correlation
One way to decide which features are going to be more impactful when running an analysis is to look at the correlation between them.
Pearson's correlations are single value numerical summaries that represent the strength and direction of a linear relationship.
Correlation values range from -1 to 1. Values further away from 0 represent stronger relationships, and the sign of the correlation (positive or negative)
represents the direction of the relationship.
The graphs below depict a visual representation of Pearson correlations.